+++ cms_helper.py
@@ -57,7 +57,7 @@ HELPER_SETTINGS = {
 
 
 def run():
-    from djangocms_helper import runner
+    from app_helper import runner
     runner.cms('djangocms_page_tags')
 
 if __name__ == "__main__":